refactor: remove unused inference utils
