--- results/america/metainfo.json
@@ -1,1 +1,0 @@
-{"data": {"fields": [{"name": "close", "type": "integer"}]}}
