--- results/bond/field_status.tsv
@@ -1,1 +1,0 @@
-field	tv_type	status	sample_value
