--- results/cfd/scan.json
@@ -1,1 +1,0 @@
-{"data": []}
