--- results/crypto/metainfo.json
@@ -1,1 +1,0 @@
-{"data": {"fields": [{"name": "close", "type": "integer"}, {"name": "open", "type": "text"}]}}
