+++ src/compare/cache_diff.py
@@ -7,14 +7,7 @@ from pathlib import Path
 from typing import Dict, Tuple
 
 
-_CUSTOM_PATTERNS = [r"^TV_Custom\.", r"_impact_score$", r"^BTC_", r"^custom_"]
-
-
-def _is_custom(name: str) -> bool:
-    for pat in _CUSTOM_PATTERNS:
-        if re.search(pat, name, re.IGNORECASE):
-            return True
-    return False
+from src.utils.custom_patterns import _CUSTOM_PATTERNS, _is_custom
 
 
 def _load_meta_fields(path: Path) -> Dict[str, str]:
