+++ src/meta/field_classifier.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 import re
 from typing import Any, Dict
 
+from src.utils.custom_patterns import _CUSTOM_PATTERNS, _is_custom
+
 # TradingView numeric types
 _NUMERIC_TYPES = {
     "number",
@@ -14,15 +16,6 @@ _NUMERIC_TYPES = {
     "duration",
     "percentage",
 }
-
-_CUSTOM_PATTERNS = [r"^TV_Custom\.", r"_impact_score$", r"^BTC_", r"^custom_"]
-
-
-def _is_custom(name: str) -> bool:
-    for pat in _CUSTOM_PATTERNS:
-        if re.search(pat, name, re.IGNORECASE):
-            return True
-    return False
 
 
 def classify_fields(columns: Dict[str, Dict[str, Any]]) -> Dict[str, list[str]]:
