+++ src/utils/__init__.py
@@ -1,6 +1,5 @@
-"""Shared utility functions for type inference."""
+"""Shared utilities for specification generation."""
 
-from .infer import infer_type
 from .type_mapping import tv2ref
 
-__all__ = ["infer_type", "tv2ref"]
+__all__ = ["tv2ref"]
